+++ lib/data/data_sub_tingkat.dart
@@ -1,0 +1,253 @@
+part of 'data.dart';
+
+final _pTeknikDasar = SubTingkat(subTingkatList: [
+  'Kuda-kuda',
+  'Sikap Pasang',
+  'Tangkisan',
+  'Pukulan',
+  'Tendangan',
+  'Elakan'
+]);
+
+final _pLatihanBeladiri = SubTingkat(subTingkatList: []);
+final _pKeindahanGerak = SubTingkat(subTingkatList: []);
+final _pKetangkasan = SubTingkat(subTingkatList: []);
+final _pOlahPernafasan = SubTingkat(subTingkatList: []);
+
+final _hitamTeknikDasar = SubTingkat(
+  subTingkatList: [
+    'Kuda-kuda',
+    'Sikap Pasang',
+    'Tangkisan',
+    'Pukulan',
+    'Tendangan',
+    'Elakan',
+  ],
+);
+
+final _hitamLatihanBeladiri = SubTingkat(subTingkatList: ['Pukul - Sambut']);
+
+final _hitamKeindahanGerak = SubTingkat(subTingkatList: ['Senam Jurus']);
+
+final _hitamKetangkasan = SubTingkat(subTingkatList: []);
+
+final _hitamOlahPernafasan = SubTingkat(
+  subTingkatList: [
+    'Dasar Power',
+    'Dasar Getaran',
+  ],
+);
+
+final _bTeknikDasar = SubTingkat(
+  subTingkatList: [
+    'Kuda - kuda',
+    'Sikap Pasang',
+    'Langkah',
+    'Tangkisan',
+    'Pukulan',
+    'Tendangan',
+    'Elakan',
+    'Sikuan',
+    'Redaman/jatuhan',
+  ],
+);
+
+final _bLatihanBeladiri = SubTingkat(
+  subTingkatList: ['Pukul - Sambut'],
+);
+
+final _bKeindahanGerak = SubTingkat(
+  subTingkatList: ['Senam Jurus'],
+);
+
+final _bKetangkasan = SubTingkat(subTingkatList: []);
+
+final _bOlahPernafasan = SubTingkat(
+  subTingkatList: [
+    'Dasar Power',
+    'Dasar Getaran',
+  ],
+);
+
+final _b1TeknikDasar = SubTingkat(
+  subTingkatList: [
+    'Sikap Pasang',
+    'Langkah',
+    'Tangkisan',
+    'Pukulan',
+    'Tendangan',
+    'Elakan',
+    'Sikuan',
+    'Redaman/jatuhan',
+    'Tangkapan Kaki',
+  ],
+);
+
+final _b1LatihanBeladiri = SubTingkat(
+  subTingkatList: [
+    'Pukul - Sambut',
+    'Jurus Pertandingan',
+    'Tuntun',
+  ],
+);
+
+final _b1KeindahanGerak = SubTingkat(subTingkatList: ['Jurus Tunggal/Solo']);
+
+final _b1Ketangkasan =
+    SubTingkat(subTingkatList: ['Sambung Pertandingan Versi IPSI']);
+
+final _b1OlahPernafasan = SubTingkat(
+  subTingkatList: [
+    'Dasar Power',
+    'Dasar Getaran',
+  ],
+);
+
+final _b2TeknikDasar = SubTingkat(subTingkatList: [
+  'Kuda - kuda',
+  'Langkah',
+  'Tangkisan',
+  'Pukulan',
+  'Tendangan',
+  'Elakan',
+  'Sikuan',
+  'Redaman',
+  'Tengkapan Kaki'
+]);
+
+final _b2LatihanBeladiri = SubTingkat(
+  subTingkatList: [
+    'Jurus Pertandingan',
+    'Tuntun',
+  ],
+);
+
+final _b2KeindahanGerak = SubTingkat(subTingkatList: []);
+
+final _b2Ketangkasan =
+    SubTingkat(subTingkatList: ['Sabung Pertandingan Versi IPSI']);
+
+final _b2OlahPernafasan = SubTingkat(
+  subTingkatList: [
+    'Dasar Power',
+    'Dasar Getaran',
+  ],
+);
+
+final _hTeknikDasar = SubTingkat(
+  subTingkatList: [
+    'Kuda - kuda',
+    'Langkah',
+    'Tangkisan',
+    'Pukulan',
+    'Tendangan',
+    'Elakan',
+    'Sikuan',
+    'Jatuhan/Bantingan',
+    'Kuncian',
+    'Redaman',
+  ],
+);
+
+final _hLatihanBeladiri = SubTingkat(
+  subTingkatList: [
+    'Jurus Pertandingan',
+    'Jurus Pembelaan',
+    'Olah Nafas',
+  ],
+);
+
+final _hKeindahanGerak = SubTingkat(
+  subTingkatList: [
+    'Jurus Tunggal/Solo',
+    'Jurus Tunggal IPSI',
+  ],
+);
+
+final _hKetangkasan =
+    SubTingkat(subTingkatList: ['Sambung Pertandingan Versi IPSI']);
+
+final _hOlahPernafasan = SubTingkat(
+  subTingkatList: [
+    'Dasar Power',
+    'Dasar Getaran',
+  ],
+);
+
+final _h1TeknikDasar = SubTingkat(
+  subTingkatList: [
+    'Kuda - kuda',
+    'Langkah',
+    'Tangkisan',
+    'Pukulan',
+    'Tendangan',
+    'Lututan',
+    'Jatuhan',
+    'Kuncian',
+    'Redaman',
+  ],
+);
+
+final _h1LatihanBeladiri = SubTingkat(
+  subTingkatList: [
+    'Jurusan Pertandingan',
+    'Jurus Pembelaan',
+    'Olah Nafas',
+  ],
+);
+
+final _h1KeindahanGerak = SubTingkat(
+  subTingkatList: [
+    'Kembangan Bebas',
+    'Jurusan Regu Persilat',
+  ],
+);
+
+final _h1Ketangkasan =
+    SubTingkat(subTingkatList: ['Sambung Pertandingan Versi IPSI']);
+
+final _h1OlahPernafasan = SubTingkat(
+  subTingkatList: [
+    'Dasar Power',
+    'Dasar Getaran',
+  ],
+);
+
+final _h2TeknikDasar = SubTingkat(
+  subTingkatList: [
+    'Kuda - kuda',
+    'Langkah',
+    'Tangkisan',
+    'Pukulan',
+    'Tendangan',
+    'Lututan',
+    'Jatuhan',
+    'Kuncian',
+    'Redaman',
+  ],
+);
+
+final _h2LatihanBeladiri = SubTingkat(
+  subTingkatList: [
+    'Jurus Pertandingan',
+    'Jurus Pembelaan',
+    'Olah Nafas',
+  ],
+);
+
+final _h2KeindahanGerak = SubTingkat(
+  subTingkatList: [
+    'Kembangan Bebas',
+    'Jurus Regu Persilat',
+  ],
+);
+
+final _h2Ketangkasan =
+    SubTingkat(subTingkatList: ['Sabung Pertandingan Versi IPSI']);
+
+final _h2OlahPernafasan = SubTingkat(
+  subTingkatList: [
+    'Dasar Power',
+    'Dasar Getaran',
+  ],
+);
