+++ lib/models/golongan.dart
@@ -1,34 +1,15 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
+
+import 'package:ksp_nusantara_app/models/tingkat.dart';
+
 class Golongan {
   final String imageUrl;
   final String name;
-  final List tingkat;
+  final List<Tingkat> tingkat;
 
   Golongan({
     required this.imageUrl,
     required this.name,
     required this.tingkat,
-  });
-}
-
-class Tingkat {
-  final String name;
-  final String form;
-  final List<String> header;
-  final List<SubTingkat> subTingkat;
-
-  Tingkat({
-    required this.name,
-    required this.form,
-    required this.header,
-    required this.subTingkat,
-  });
-}
-
-class SubTingkat {
-  final List<String> subTingkatList;
-
-  SubTingkat({
-    required this.subTingkatList,
   });
 }
