+++ lib/models/sub_tingkat.dart
@@ -1,0 +1,7 @@
+class SubTingkat {
+  final List<String> subTingkatList;
+
+  SubTingkat({
+    required this.subTingkatList,
+  });
+}
