+++ lib/screen/final_materi_screen.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:native_video_view/native_video_view.dart';
 
 class FinalMateriScreen extends StatelessWidget {
   const FinalMateriScreen({super.key, required this.title});
@@ -11,12 +12,23 @@ class FinalMateriScreen extends StatelessWidget {
       appBar: AppBar(
         title: Text(title),
       ),
-      body: SingleChildScrollView(
-        child: Container(
-          padding: const EdgeInsets.all(10.0),
-          child: Column(
-            children: const [],
-          ),
+      body: Container(
+        child: NativeVideoView(
+          onCreated: (controller) {
+            controller.setVideoSource('assets/video/test.mp4');
+          },
+          onPrepared: (controller, info) {
+            controller.play();
+          },
+          onCompletion: (controller) {
+            print('Video completed');
+          },
+          onError: (controller, what, extra, message) {
+            print('Player error ($what | $extra | $message)');
+          },
+          onProgress: (progress, duration) {
+            print('$progress | $duration');
+          },
         ),
       ),
     );
